added issuer APIs (skeletons)
