+++ e2e-tests/bom-tests/src/test/java/org/eclipse/edc/test/bom/BomSmokeTests.java
@@ -124,8 +124,8 @@ public class BomSmokeTests {
                                 put("web.http.version.path", "/api/version");
                                 put("web.http.did.port", valueOf(getFreePort()));
                                 put("web.http.did.path", "/api/did");
-                                put("web.http.credential-request.port", valueOf(getFreePort()));
-                                put("web.http.credential-request.path", "/api/issuance");
+                                put("web.http.issuer-api.port", valueOf(getFreePort()));
+                                put("web.http.issuer-api.path", "/api/issuer");
                                 put("edc.sts.account.api.url", "https://sts.com/accounts");
                                 put("edc.sts.accounts.api.auth.header.value", "password");
                             }
