--- extensions/protocols/dcp/credential-request-api/src/main/java/org/eclipse/edc/protocols/dcp/credentialrequest/v1alpha/api/CredentialRequestApiController.java
@@ -1,37 +1,0 @@
-/*
- *  Copyright (c) 2025 Cofinity-X
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Cofinity-X - initial API and implementation
- *
- */
-
-package org.eclipse.edc.protocols.dcp.credentialrequest.v1alpha.api;
-
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.Response;
-import org.eclipse.edc.protocols.dcp.credentialrequest.v1alpha.model.CredentialRequestMessage;
-
-import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
-
-@Consumes(APPLICATION_JSON)
-@Produces(APPLICATION_JSON)
-@Path("/v1alpha/credentials")
-public class CredentialRequestApiController implements CredentialRequestApi {
-
-    @POST
-    @Path("/")
-    @Override
-    public Response requestCredential(CredentialRequestMessage message) {
-        return null;
-    }
-}
