--- extensions/protocols/dcp/credential-request-api/src/main/java/org/eclipse/edc/protocols/dcp/credentialrequest/v1alpha/model/CredentialRequest.java
@@ -1,22 +1,0 @@
-/*
- *  Copyright (c) 2025 Cofinity-X
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Cofinity-X - initial API and implementation
- *
- */
-
-package org.eclipse.edc.protocols.dcp.credentialrequest.v1alpha.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record CredentialRequest(@JsonProperty("credentialType") String credentialType,
-                                @JsonProperty("format") String format,
-                                @JsonProperty("payload") Object payload) {
-}
