--- extensions/protocols/dcp/credential-request-api/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,15 +1,0 @@
-#
-#  Copyright (c) 2025 Cofinity-X
-#
-#  This program and the accompanying materials are made available under the
-#  terms of the Apache License, Version 2.0 which is available at
-#  https://www.apache.org/licenses/LICENSE-2.0
-#
-#  SPDX-License-Identifier: Apache-2.0
-#
-#  Contributors:
-#       Cofinity-X - initial API and implementation
-#
-#
-
-org.eclipse.edc.protocols.dcp.credentialrequest.CredentialRequestApiExtension
