--- extensions/protocols/dcp/credential-request-status-api/src/main/resources/credential-request-status-api-version.json
@@ -1,8 +1,0 @@
-[
-  {
-    "version": "1.0.0",
-    "urlPath": "/v1alpha",
-    "lastUpdated": "2025-01-16T12:00:00Z",
-    "maturity": null
-  }
-]
