--- resources/openapi/credential-request-api.version
@@ -1,1 +1,0 @@
-extensions/protocols/dcp/credential-request-api/src/main/resources/credential-request-api-version.json
