--- resources/openapi/credential-request-status-api.version
@@ -1,1 +1,0 @@
-extensions/protocols/dcp/credential-request-status-api/src/main/resources/credential-request-status-api-version.json
