+++ resources/openapi/issuer-api.version
@@ -1,0 +1,1 @@
+extensions/protocols/dcp/credential-request-api/src/main/resources/issuer-api-version.json
