--- resources/openapi/issuer-metadata-api.version
@@ -1,1 +1,0 @@
-extensions/protocols/dcp/issuer-metadata-api/src/main/resources/issuer-metadata-api-version.json
