+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/webcontext/IdentityHubApiContext.java
@@ -18,7 +18,7 @@ public interface IdentityHubApiContext {
     String IDENTITY = "identity";
     String IH_DID = "did";
     String PRESENTATION = "presentation";
-    String CREDENTIAL_REQUEST = "credential-request";
+    String ISSUER_API = "issuer-api";
     @Deprecated(since = "0.9.0")
     String RESOLUTION = "resolution";
 }
